Moved all code in rs/tests/src/nns_tests/bitcoin_set_config_proposal.rs to rs/tests/nns/bitcoin_set_config_proposal_test.rs (an existing file). This way, the bitcoin_set_config_proposal_test does not need to depend on //rs/tests. Thus, a random change within //rs/tests does not invalidate runs of this test.
