+++ rs/tests/nns/Cargo.toml
@@ -7,17 +7,24 @@ description.workspace = true
 documentation.workspace = true
 
 [dependencies]
-anyhow = { workspace = true }
-ic-mainnet-nns-recovery = { path = "ic_mainnet_nns_recovery" }
-ic-registry-subnet-type = { path = "../../registry/subnet_type" }
-ic-system-test-driver = { path = "../driver" }
-slog = { workspace = true }
-tests = { path = ".." }
-candid = { workspace = true }
-ic_consensus_system_test_utils = { path = "../consensus/utils" }
-nns_dapp = { path = "nns_dapp" }
-libflate = { workspace = true }
-reqwest = { workspace = true }
+canister-test                   = { path = "../../rust_canisters/canister_test" }
+ic-config                       = { path = "../../config" }
+ic-nns-constants                = { path = "../../nns/constants" }
+ic-nns-governance-api           = { path = "../../nns/governance/api" }
+ic-nns-test-utils               = { path = "../../nns/test_utils" }
+ic-registry-subnet-type         = { path = "../../registry/subnet_type" }
+ic-system-test-driver           = { path = "../../tests/driver" }
+ic-tests-ckbtc                  = { path = "../../tests/ckbtc" }
+ic_consensus_system_test_utils  = { path = "../../tests/consensus/utils" }
+nns_dapp                        = { path = "../../tests/nns/nns_dapp" }
+
+anyhow                          = { workspace = true }
+candid                          = { workspace = true }
+ic-agent                        = { workspace = true }
+ic-btc-interface                = { workspace = true }
+libflate                        = { workspace = true }
+reqwest                         = { workspace = true }
+slog                            = { workspace = true }
 
 [[bin]]
 name = "nns_upgrade_test"
@@ -48,13 +55,13 @@ name = "nns_token_balance_test"
 path = "nns_token_balance_test.rs"
 
 [[bin]]
-name = "nns_upgrade_test"
-path = "nns_upgrade_test.rs"
-
-[[bin]]
 name = "node_removal_from_registry_test"
 path = "node_removal_from_registry_test.rs"
 
 [[bin]]
 name = "nns_dapp_test"
 path = "nns_dapp_test.rs"
+
+[[bin]]
+name = "bitcoin-set-config-proposal-test"
+path = "bitcoin_set_config_proposal_test.rs"
