+++ rs/tests/src/nns_tests/mod.rs
@@ -1,4 +1,3 @@
-pub mod bitcoin_set_config_proposal;
 pub mod create_subnet;
 pub mod cycles_minting;
 pub mod neurons_fund;
